LayoutTests/imported/w3c/web-platform-tests/referrer-policy/gen/top.http-rp/origin/xhr.http-expected.txt

Blocked access to external URL http://www1.localhost:8800/common/security-features/subresource/xhr.py?redirection=keep-origin&action=purge&key=7f852a25-3d88-486e-8298-18734af08b6a&path=%2Fmixed-content
CONSOLE MESSAGE: XMLHttpRequest cannot load http://www1.localhost:8800/common/security-features/subresource/xhr.py?redirection=keep-origin&action=purge&key=7f852a25-3d88-486e-8298-18734af08b6a&path=%2Fmixed-content due to access control checks.
Blocked access to external URL http://www1.localhost:8800/common/security-features/subresource/xhr.py?redirection=no-redirect&action=purge&key=9b0daa2d-8629-4d64-9a57-975eedae4657&path=%2Fmixed-content
CONSOLE MESSAGE: XMLHttpRequest cannot load http://www1.localhost:8800/common/security-features/subresource/xhr.py?redirection=no-redirect&action=purge&key=9b0daa2d-8629-4d64-9a57-975eedae4657&path=%2Fmixed-content due to access control checks.
Blocked access to external URL http://www1.localhost:8800/common/security-features/subresource/xhr.py?redirection=swap-origin&action=purge&key=80f6a245-db2e-48d4-8289-fcbbf1bcfe14&path=%2Fmixed-content
CONSOLE MESSAGE: XMLHttpRequest cannot load http://www1.localhost:8800/common/security-features/subresource/xhr.py?redirection=swap-origin&action=purge&key=80f6a245-db2e-48d4-8289-fcbbf1bcfe14&path=%2Fmixed-content due to access control checks.
Blocked access to external URL https://www1.localhost:9443/common/security-features/subresource/xhr.py?redirection=keep-origin&action=purge&key=2f856726-9b8c-4487-997f-1c25c176a6c3&path=%2Fmixed-content
CONSOLE MESSAGE: XMLHttpRequest cannot load https://www1.localhost:9443/common/security-features/subresource/xhr.py?redirection=keep-origin&action=purge&key=2f856726-9b8c-4487-997f-1c25c176a6c3&path=%2Fmixed-content due to access control checks.
Blocked access to external URL https://www1.localhost:9443/common/security-features/subresource/xhr.py?redirection=no-redirect&action=purge&key=3c5b7be4-c154-4771-9d11-28d9a88e72f6&path=%2Fmixed-content
CONSOLE MESSAGE: XMLHttpRequest cannot load https://www1.localhost:9443/common/security-features/subresource/xhr.py?redirection=no-redirect&action=purge&key=3c5b7be4-c154-4771-9d11-28d9a88e72f6&path=%2Fmixed-content due to access control checks.
Blocked access to external URL https://www1.localhost:9443/common/security-features/subresource/xhr.py?redirection=swap-origin&action=purge&key=8a29effd-bf00-4814-a7c1-e9291e28a1b5&path=%2Fmixed-content
CONSOLE MESSAGE: XMLHttpRequest cannot load https://www1.localhost:9443/common/security-features/subresource/xhr.py?redirection=swap-origin&action=purge&key=8a29effd-bf00-4814-a7c1-e9291e28a1b5&path=%2Fmixed-content due to access control checks.
Blocked access to external URL http://www1.localhost:8800/common/security-features/subresource/xhr.py?action=purge&key=b1e153fd-1fc2-432d-b7ef-a50f6962c5d6&path=%2Fmixed-content
Blocked access to external URL https://www1.localhost:9443/common/security-features/subresource/xhr.py?action=purge&key=09f8f913-dff7-48bb-87fd-73926b76c17a&path=%2Fmixed-content

FAIL Referrer Policy: Expects origin for xhr to cross-http origin and keep-origin redirection from http context. promise_test: Unhandled rejection with value: object "Error: Network Error"
FAIL Referrer Policy: Expects origin for xhr to cross-http origin and no-redirect redirection from http context. promise_test: Unhandled rejection with value: object "Error: Network Error"
FAIL Referrer Policy: Expects origin for xhr to cross-http origin and swap-origin redirection from http context. promise_test: Unhandled rejection with value: object "Error: Network Error"
FAIL Referrer Policy: Expects origin for xhr to cross-https origin and keep-origin redirection from http context. promise_test: Unhandled rejection with value: object "Error: Network Error"
FAIL Referrer Policy: Expects origin for xhr to cross-https origin and no-redirect redirection from http context. promise_test: Unhandled rejection with value: object "Error: Network Error"
FAIL Referrer Policy: Expects origin for xhr to cross-https origin and swap-origin redirection from http context. promise_test: Unhandled rejection with value: object "Error: Network Error"
PASS Referrer Policy: Expects origin for xhr to same-http origin and keep-origin redirection from http context.
PASS Referrer Policy: Expects origin for xhr to same-http origin and no-redirect redirection from http context.
FAIL Referrer Policy: Expects origin for xhr to same-http origin and swap-origin redirection from http context. promise_test: Unhandled rejection with value: object "Error: Network Error"
PASS Referrer Policy: Expects origin for xhr to same-https origin and keep-origin redirection from http context.
PASS Referrer Policy: Expects origin for xhr to same-https origin and no-redirect redirection from http context.
FAIL Referrer Policy: Expects origin for xhr to same-https origin and swap-origin redirection from http context. promise_test: Unhandled rejection with value: object "Error: Network Error"
